SASS-4445: House keeping
Add @Singleton to controllers and rename Tests with Test suffix to Spec suffix
