+++ app/controllers/PensionChargesController.scala
@@ -26,9 +26,10 @@ import services.{PensionChargesService, PensionsService}
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
 
-import javax.inject.Inject
+import javax.inject.{Inject, Singleton}
 import scala.concurrent.{ExecutionContext, Future}
 
+@Singleton
 class PensionChargesController @Inject()(service: PensionChargesService,
                                          authorisedAction: AuthorisedAction,
                                          cc: ControllerComponents,
