+++ it/api/DeletePensionChargesISpec.scala
@@ -17,38 +17,14 @@
 package api
 
 import helpers.WiremockSpec
-import models.{CreateOrUpdatePensionReliefsModel, DesErrorBodyModel, PensionReliefs}
+import models.DesErrorBodyModel
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.time.{Seconds, Span}
 import play.api.http.Status._
-import play.api.libs.json.{JsValue, Json}
+import play.api.libs.json.Json
 import utils.DESTaxYearHelper.desTaxYearConverter
 
-class CreateOrAmendPensionReliefsITest extends WiremockSpec with ScalaFutures {
-
-  val minimumPensionReliefs: PensionReliefs = PensionReliefs(
-    regularPensionContributions = Some(10.22), None, None, None, None)
-
-  val fullPensionReliefs: PensionReliefs = PensionReliefs(
-    regularPensionContributions = Some(10.22),
-    oneOffPensionContributionsPaid = Some(11.33),
-    retirementAnnuityPayments = Some(12.44),
-    paymentToEmployersSchemeNoTaxRelief = Some(13.55),
-    overseasPensionSchemeContributions = Some(14.66))
-
-  val fullCreateOrUpdatePensionReliefsData: CreateOrUpdatePensionReliefsModel = CreateOrUpdatePensionReliefsModel(fullPensionReliefs)
-  val minEmploymentFinancialData: CreateOrUpdatePensionReliefsModel = CreateOrUpdatePensionReliefsModel(minimumPensionReliefs)
-
-  val fullJson: JsValue = Json.parse(
-    """{
-      |	"pensionReliefs": {
-      |		"regularPensionContributions": 10.22,
-      |		"oneOffPensionContributionsPaid": 11.33,
-      |		"retirementAnnuityPayments": 12.44,
-      |		"paymentToEmployersSchemeNoTaxRelief": 13.55,
-      |		"overseasPensionSchemeContributions": 14.66
-      |	}
-      |}""".stripMargin)
+class DeletePensionChargesISpec extends WiremockSpec with ScalaFutures {
 
   trait Setup {
     val timeSpan: Long = 5
@@ -58,111 +34,136 @@ class CreateOrAmendPensionReliefsITest extends WiremockSpec with ScalaFutures {
     val mtditidHeader: (String, String) = ("mtditid", "555555555")
     val mtdBearerToken : (String, String) = ("Authorization", "Bearer:XYZ")
     val requestHeaders: Seq[(String, String)] = Seq(mtditidHeader, mtdBearerToken)
-    val desUrl = s"/income-tax/reliefs/pensions/$nino/${desTaxYearConverter(taxYear)}"
-    val serviceUrl: String = s"/income-tax-pensions/pension-reliefs/nino/$nino/taxYear/$taxYear"
+    val desUrl: String = s"/income-tax/charges/pensions/$nino/${desTaxYearConverter(taxYear)}"
+    val serviceUrl: String = s"/income-tax-pensions/pension-charges/nino/$nino/taxYear/$taxYear"
     auditStubs()
   }
 
-  "create or amend pension reliefs" when {
+  "delete pension charges" when {
 
     "the user is an individual" must {
       "return a No content Success response" in new Setup {
 
-        stubPutWithoutResponseBody(desUrl, Json.toJson(fullCreateOrUpdatePensionReliefsData).toString(), NO_CONTENT)
+        stubDeleteWithoutResponseBody(desUrl, NO_CONTENT)
 
         authorised()
 
         whenReady(buildClient(serviceUrl)
           .withHttpHeaders(requestHeaders:_*)
-          .put(fullJson)) {
+          .delete()) {
           result =>
             result.status mustBe NO_CONTENT
 
         }
       }
 
-      "return 400 if the body payload validation fails" in new Setup {
+      "return 400 if a there is an invalid tax year" in new Setup {
+
+        val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
+          "INVALID_TAX_YEAR", "Submission has not passed validation. Invalid parameter taxYear.")).toString()
+
+        stubDeleteWithResponseBody(desUrl, BAD_REQUEST, errorResponseBody)
 
         authorised()
 
         whenReady(buildClient(serviceUrl)
           .withHttpHeaders(requestHeaders:_*)
-          .put(Json.obj())) {
+          .delete()) {
           result =>
             result.status mustBe BAD_REQUEST
+            Json.parse(result.body) mustBe Json.obj(
+              "code" -> "INVALID_TAX_YEAR", "reason" -> "Submission has not passed validation. Invalid parameter taxYear.")
         }
       }
 
-      "return 400 if a downstream bad request error occurs" in new Setup {
+      "return 400 if a there is an invalid taxable entity id (nino)" in new Setup {
 
         val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
           "INVALID_TAXABLE_ENTITY_ID", "Submission has not passed validation. Invalid parameter taxableEntityId.")).toString()
 
-        stubPutWithResponseBody(desUrl, Json.toJson(fullCreateOrUpdatePensionReliefsData).toString(), errorResponseBody, BAD_REQUEST)
+        stubDeleteWithResponseBody(desUrl, BAD_REQUEST, errorResponseBody)
 
         authorised()
 
         whenReady(buildClient(serviceUrl)
           .withHttpHeaders(requestHeaders:_*)
-          .put(fullJson)) {
+          .delete()) {
           result =>
             result.status mustBe BAD_REQUEST
-            result.body mustBe errorResponseBody
+            Json.parse(result.body) mustBe Json.obj(
+              "code" -> "INVALID_TAXABLE_ENTITY_ID", "reason" -> "Submission has not passed validation. Invalid parameter taxableEntityId.")
         }
       }
 
-      "return 500 if an unexpected error is returned from DES user" in new Setup {
+      "return 400 if a there is an invalid header correlation id" in new Setup {
 
-        // e,g, 404 not found is not expected as create or update will create if not found
-        val errorResponseBody: String = Json.toJson(DesErrorBodyModel.parsingError).toString()
+        val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
+          "INVALID_CORRELATIONID", "Submission has not passed validation. Invalid Header parameter CorrelationId.")).toString()
 
-        stubPutWithResponseBody(desUrl, Json.toJson(fullCreateOrUpdatePensionReliefsData).toString(), errorResponseBody, NOT_FOUND)
+        stubDeleteWithResponseBody(desUrl, BAD_REQUEST, errorResponseBody)
 
         authorised()
 
         whenReady(buildClient(serviceUrl)
           .withHttpHeaders(requestHeaders:_*)
-          .put(fullJson)) {
+          .delete()) {
           result =>
-            result.status mustBe INTERNAL_SERVER_ERROR
+            result.status mustBe BAD_REQUEST
+            Json.parse(result.body) mustBe Json.obj(
+              "code" -> "INVALID_CORRELATIONID", "reason" -> "Submission has not passed validation. Invalid Header parameter CorrelationId.")
+        }
+      }
+
+      "return 404 if a user has no recorded pension charges to delete" in new Setup {
+
+        val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
+          "NO_DATA_FOUND", "The remote endpoint has indicated that no data can be found.")).toString()
+
+        stubDeleteWithResponseBody(desUrl, NOT_FOUND, errorResponseBody)
+
+        authorised()
+
+        whenReady(buildClient(serviceUrl)
+          .withHttpHeaders(requestHeaders:_*)
+          .delete()) {
+          result =>
+            result.status mustBe NOT_FOUND
             result.body mustBe errorResponseBody
         }
       }
 
-      "return 503 if a downstream service unavailable error occurs" in new Setup {
+      "return 503 if a downstream error occurs" in new Setup {
 
         val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
           "SERVICE_UNAVAILABLE", "Dependent systems are currently not responding.")).toString()
 
-        stubPutWithResponseBody(desUrl, Json.toJson(fullCreateOrUpdatePensionReliefsData).toString(), errorResponseBody, SERVICE_UNAVAILABLE)
+        stubDeleteWithResponseBody(desUrl, SERVICE_UNAVAILABLE, errorResponseBody)
 
         authorised()
 
         whenReady(buildClient(serviceUrl)
           .withHttpHeaders(requestHeaders:_*)
-          .put(fullJson)) {
+          .delete()) {
           result =>
             result.status mustBe SERVICE_UNAVAILABLE
-            result.body mustBe errorResponseBody
             Json.parse(result.body) mustBe Json.obj("code" -> "SERVICE_UNAVAILABLE", "reason" -> "Dependent systems are currently not responding.")
         }
       }
 
-      "return 500 if a downstream internal server error occurs" in new Setup {
+      "return 500 if a downstream error occurs" in new Setup {
 
         val errorResponseBody: String = Json.toJson(DesErrorBodyModel(
           "SERVER_ERROR", "DES is currently experiencing problems that require live service intervention.")).toString()
 
-        stubPutWithResponseBody(desUrl, Json.toJson(fullCreateOrUpdatePensionReliefsData).toString(), errorResponseBody, INTERNAL_SERVER_ERROR)
+        stubDeleteWithResponseBody(desUrl, INTERNAL_SERVER_ERROR, errorResponseBody)
 
         authorised()
 
         whenReady(buildClient(serviceUrl)
           .withHttpHeaders(requestHeaders:_*)
-          .put(fullJson)) {
+          .delete()) {
           result =>
             result.status mustBe INTERNAL_SERVER_ERROR
-            result.body mustBe errorResponseBody
             Json.parse(result.body) mustBe Json.obj(
               "code" -> "SERVER_ERROR", "reason" -> "DES is currently experiencing problems that require live service intervention.")
         }
@@ -173,7 +174,7 @@ class CreateOrAmendPensionReliefsITest extends WiremockSpec with ScalaFutures {
 
         whenReady(buildClient(serviceUrl)
           .withHttpHeaders(requestHeaders:_*)
-          .put(fullJson)) {
+          .delete()) {
           result =>
             result.status mustBe UNAUTHORIZED
             result.body mustBe ""
@@ -182,7 +183,7 @@ class CreateOrAmendPensionReliefsITest extends WiremockSpec with ScalaFutures {
 
       "return 401 if the request has no MTDITID header present" in new Setup {
         whenReady(buildClient(serviceUrl)
-          .put(fullJson)) {
+          .delete()) {
           result =>
             result.status mustBe UNAUTHORIZED
             result.body mustBe ""
